# Refactor: unpack and re-order HTMLParser.mainLoop foreign-content conditionals