This changeset should functionally be a no-op, and may be best reviewed with whitespace changes ignored since the indentation level for some of the loop's logic has been reduced.

The changes appear to result in a minor performance improvement; that said there are likely larger benefits to be found by further simplification and refactoring (perhaps including fairly substantial logical restructuring).

**Before** (2c19b9899ab3a3e8bd0ca35e5d78544334204169)
```
.........................................
html_parse_etree: Mean +- std dev: 204 ms +- 10 ms
```

**After** (369a4123ae453845a383dafb82e7459a4e87839b)
```
.........................................
html_parse_etree: Mean +- std dev: 197 ms +- 9 ms
```